pytest: Separate report tests
